Strip prefix from enum values
It is best practice in protobuf to prefix all enum values with the enum's name
to avoid conflicts in languages where all enum values are in the global
namespace.

Most protobuf compilers for languages where enum values are not in the global
namespace strip this prefix to avoid the redundant name duplication.

Example:

enum InternalColorId {
  INTERNAL_COLOR_ID_UNSPECIFIED = 0;
  INTERNAL_COLOR_ID_RED = 1;
  INTERNAL_COLOR_ID_GREEN = 2;
}

becomes

public enum InternalColorId {
  Unspecified = 0,
  Red = 1,
  Green = 2
}

in C#.

This change makes it so that Godobuf strips the prefix as well,
allowing you to write

var color := InternalColorId.RED;

instead of

var color := InternalColorId.INTERNAL_COLOR_ID_RED;

The prefix is stripped only if it matches the enum name and if every value
in the enum uses the same prefix.