Fix incorrect type_class_id's after merging of imported files
When generating GDScript code from protobuf file, that uses imports, that are using imports themselves, it can occur, that after merging the analysis data from multiple files the type_class_id values will get invalidated. Because of that, they need to be recalculated.

This change fixes a bug, that occur when after the second run of semantock.check() indexes are not getting reassigned, because the program keeps the result of previous run. Because of that bug, the generated code becomes completely invalid.